Added chunk sub allocations to GPU page Refactor memory reporting Added instance hash to framebuffer to support resizing

+++ GraphicsEngine/Source/D3D12RHI/RHI/RenderAPIs/D3D12/D3D12Helpers.h
@@ -40,7 +40,7 @@ public:
 	static D3D12_RESOURCE_DIMENSION ConvertToResourceDimension(eTextureDimension Dim);
 	static D3D12_COMMAND_LIST_TYPE ConvertListType(ECommandListType::Type type);
 	static D3D12_RESOURCE_STATES ConvertBufferResourceState(EBufferResourceState::Type intype);
-	static UINT Align(UINT size, UINT alignment = D3D12_DEFAULT_RESOURCE_PLACEMENT_ALIGNMENT)
+	static UINT64 Align(UINT64 size, UINT64 alignment = D3D12_DEFAULT_RESOURCE_PLACEMENT_ALIGNMENT)
 	{
 		return (size + alignment - 1) & ~(alignment - 1);
 	}
